third_party/blink/web_tests/svg/animations/animate-elem-05-t-drt-expected.txt

SVG 1.1 dynamic animation tests

A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getTransformToElement(rootSVGElement, path).e is -15
PASS getTransformToElement(rootSVGElement, path).f is -43
PASS getTransformToElement(rootSVGElement, path).e is -40
PASS getTransformToElement(rootSVGElement, path).f is -30
PASS getTransformToElement(rootSVGElement, path).e is -65
PASS getTransformToElement(rootSVGElement, path).f is -30
PASS getTransformToElement(rootSVGElement, path).e is -65
PASS getTransformToElement(rootSVGElement, path).f is -30
PASS successfullyParsed is true

TEST COMPLETE
